requirements.txt
black==22.3.0
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.1.1
zipp==3.8.0
